Move Ruby plugin to plugin package and refactor checklist to provide language specific lists.
