+++ sonar-kotlin-plugin/src/main/java/org/sonarsource/slang/kotlin/plugin/KotlinRulesDefinition.java
@@ -36,7 +36,7 @@ public class KotlinRulesDefinition implements RulesDefinition {
       .setName(KotlinPlugin.REPOSITORY_NAME);
     RuleMetadataLoader ruleMetadataLoader = new RuleMetadataLoader(RESOURCE_FOLDER);
 // TODO: Add logic for rules that require language specific configuration at construction time
-    ArrayList<Class> checks = new ArrayList<>(CommonCheckList.checks());
+    ArrayList<Class> checks = new ArrayList<>(CommonCheckList.kotlinChecks());
     checks.add(CommentedCodeCheck.class);
     ruleMetadataLoader.addRulesByAnnotatedClass(repository, checks);
     repository.done();
