+++ sonar-ruby-plugin/pom.xml
@@ -26,6 +26,10 @@
             <groupId>org.sonarsource.analyzer-commons</groupId>
             <artifactId>sonar-analyzer-commons</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.sonarsource.slang</groupId>
+            <artifactId>slang-checks</artifactId>
+        </dependency>
 
         <dependency>
             <groupId>junit</groupId>
@@ -46,7 +50,7 @@
                 <configuration>
                     <!-- This value shows up in the SonarQube GUI -->
                     <pluginName>SonarRuby</pluginName>
-                    <pluginClass>org.sonarsource.slang.ruby.RubyPlugin</pluginClass>
+                    <pluginClass>org.sonarsource.slang.ruby.plugin.RubyPlugin</pluginClass>
                     <skipDependenciesPackaging>true</skipDependenciesPackaging>
                     <sonarLintSupported>false</sonarLintSupported>
                     <addMavenDescriptor>false</addMavenDescriptor>
@@ -90,8 +94,8 @@
                         <configuration>
                             <rules>
                                 <requireFilesSize>
-                                    <minsize>40000</minsize>
-                                    <maxsize>60000</maxsize>
+                                    <minsize>100000</minsize>
+                                    <maxsize>300000</maxsize>
                                     <files>
                                         <file>${project.build.directory}/${project.build.finalName}.jar</file>
                                     </files>
