--- sonar-ruby-plugin/src/main/java/org/sonarsource/slang/ruby/RubyRulesDefinition.java
@@ -1,43 +1,0 @@
-/*
- * SonarSource SLang
- * Copyright (C) 2009-2018 SonarSource SA
- * mailto:info AT sonarsource DOT com
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- */
-package org.sonarsource.slang.ruby;
-
-import java.util.List;
-import org.sonar.api.server.rule.RulesDefinition;
-import org.sonarsource.analyzer.commons.RuleMetadataLoader;
-
-public class RubyRulesDefinition implements RulesDefinition {
-
-  private static final String RESOURCE_FOLDER = "org/sonar/l10n/ruby/rules/ruby";
-
-  @Override
-  public void define(Context context) {
-    NewRepository repository = context
-      .createRepository(RubyPlugin.RUBY_REPOSITORY_KEY, RubyPlugin.RUBY_LANGUAGE_KEY)
-      .setName(RubyPlugin.REPOSITORY_NAME);
-    RuleMetadataLoader ruleMetadataLoader = new RuleMetadataLoader(RESOURCE_FOLDER);
-
-
-    List<Class> checks = RubyCheckList.checks();
-    ruleMetadataLoader.addRulesByAnnotatedClass(repository, checks);
-    repository.done();
-  }
-
-}
