+++ sonar-ruby-plugin/src/main/java/org/sonarsource/slang/ruby/plugin/RubyLanguage.java
@@ -17,5 +17,27 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-@javax.annotation.ParametersAreNonnullByDefault
-package org.sonarsource.slang.ruby;
+package org.sonarsource.slang.ruby.plugin;
+
+import org.sonar.api.config.Configuration;
+import org.sonar.api.resources.AbstractLanguage;
+
+public class RubyLanguage extends AbstractLanguage {
+
+  private Configuration configuration;
+
+  public RubyLanguage(Configuration configuration) {
+    super(RubyPlugin.RUBY_LANGUAGE_KEY, RubyPlugin.RUBY_LANGUAGE_NAME);
+    this.configuration = configuration;
+  }
+
+  @Override
+  public String[] getFileSuffixes() {
+    String[] suffixes = configuration.getStringArray(RubyPlugin.RUBY_FILE_SUFFIXES_KEY);
+    if (suffixes == null || suffixes.length == 0) {
+      suffixes = RubyPlugin.RUBY_FILE_SUFFIXES_DEFAULT_VALUE.split(",");
+    }
+    return suffixes;
+  }
+
+}
