+++ sonar-ruby-plugin/src/main/java/org/sonarsource/slang/ruby/plugin/RubyPlugin.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.slang.ruby;
+package org.sonarsource.slang.ruby.plugin;
 
 import org.sonar.api.Plugin;
 import org.sonar.api.config.PropertyDefinition;
