+++ sonar-ruby-plugin/src/main/java/org/sonarsource/slang/ruby/plugin/RubyProfileDefinition.java
@@ -17,17 +17,20 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.slang.ruby;
+package org.sonarsource.slang.ruby.plugin;
 
-import java.util.Arrays;
-import java.util.List;
+import org.sonar.api.server.profile.BuiltInQualityProfilesDefinition;
+import org.sonarsource.analyzer.commons.BuiltInQualityProfileJsonLoader;
 
-class RubyCheckList {
+public class RubyProfileDefinition implements BuiltInQualityProfilesDefinition {
 
-  private RubyCheckList() {
+  private static final String PATH_TO_JSON = "org/sonar/l10n/ruby/rules/ruby/Sonar_way_profile.json";
+
+  @Override
+  public void define(Context context) {
+    NewBuiltInQualityProfile profile = context.createBuiltInQualityProfile(RubyPlugin.PROFILE_NAME, RubyPlugin.RUBY_LANGUAGE_KEY);
+    BuiltInQualityProfileJsonLoader.load(profile, RubyPlugin.RUBY_REPOSITORY_KEY, PATH_TO_JSON);
+    profile.done();
   }
 
-  static List<Class> checks() {
-    return Arrays.asList();
-  }
 }
